refactor: Extract SyncManager service
