+++ src/content/services/index.ts
@@ -1,3 +1,4 @@
 export type { Service } from './service';
 
+export { default as SyncManager } from './sync-manager';
 export { default as UIManager } from './ui-manager';
